backend/requirements.txt
flask
python-dotenv
flask-restful
flask_sqlalchemy
flask-jwt-extended
flask-mail
flask-login
SQLAlchemy
marshmallow>=3.0.0